+++ client/test/extension.test.ts
@@ -12,11 +12,14 @@ import * as vscode from 'vscode';
 import * as myExtension from '../src/extension';
 
 // Defines a Mocha test suite to group tests of similar kind together
-suite("Extension Tests", () => {
+suite("UI5 Typescript Tool Extension Tests", () => {
 
-    // Defines a Mocha unit test
-    test("Something 1", () => {
-        assert.equal(-1, [1, 2, 3].indexOf(5));
-        assert.equal(-1, [1, 2, 3].indexOf(0));
-    });
+    test("Check function getManifestLocation", async() => {
+        await myExtension.getManifestLocation();
+        assert.notEqual(myExtension.ui5tsglobal.core.absoluteRootPath.endsWith, "client-test\\webapp");
+        assert.equal(myExtension.ui5tsglobal.core.relativeRootPath, "webapp");
+
+    }); 
+ 
+
 });